Add windows support

+++ include/pktbuilder/UDP.h
@@ -13,8 +13,8 @@ namespace pktbuilder {
             bool no_checksum = false;
         public:
             Datagram() = default;
-            explicit Datagram(uint32_t destination_port, uint32_t source_port = 0, bool no_checksum = false);
-            Datagram(ipv4_addr_t destination_address, uint32_t destination_port, ipv4_addr_t source_address, uint32_t source_port = 0);
+            explicit Datagram(uint16_t destination_port, uint16_t source_port = 0, bool no_checksum = false);
+            Datagram(ipv4_addr_t destination_address, uint16_t destination_port, ipv4_addr_t source_address, uint16_t source_port = 0);
             IPv4::Packet operator|(IPv4::Packet const& other);
             [[nodiscard]] std::vector<uint8_t> build() const override;
         };
